prototype for EndDungeonController
DBConstants.java - user task history
Event.proto - end dungeon request has clientTime; response has FullUserEquipProto
Info.proto - userEquipId in FullUserEquipProto is a bigint
UserTaskRetrieveUtils.java - userTaskId is a bigint
BeginDungeonController.java - mistakenly gave user exp and silver and recorded that user completed task
CreateInfoProtoUtils.java - creating FullUserEquipProto with userEquipId a bigint/long
InsertUtils.java - user_task_history table

+++ src/main/java/com/lvl6/retrieveutils/UserTaskRetrieveUtils.java
@@ -26,7 +26,7 @@ import com.lvl6.utils.DBConnection;
   
   private static final String TABLE_NAME = DBConstants.TABLE_USER_TASK;
   
-  public static UserTask getUserTaskForId(int userTaskId) {
+  public static UserTask getUserTaskForId(long userTaskId) {
     Connection conn = DBConnection.get().getConnection();
     Map<String, Object> absoluteConditionParams = new HashMap<String, Object>();
     absoluteConditionParams.put(DBConstants.USER_TASK__ID, userTaskId);
